requirements.txt
python:
http.server

pip:
speedtest-cli
pandas

commands:
python3 -m http.server
python3 main.py


Run using crontab
Edit 5 to however many minutes between test

*/5 * * * * cd /path/to/file && python3 main.py
@reboot cd /path/to/file && python3 -m http.server &
